Since this platform is for skillz, I decided to redesign the types table and store more of the skill information specific to individual skill in their respective row in the types table, thus reducing the processing time creating arrays and building collections. Lines were delete and code was clean up as a result of this. :-D The ability to import a list of skillz into the plan as been enabled. This is going to be fun. Added an import implants method to the setup process :-D

+++ database/migrations/2018_06_27_023105_create_type_dogma_attributes_table.php
@@ -16,7 +16,7 @@ class CreateTypeDogmaAttributesTable extends Migration
         Schema::create('type_dogma_attributes', function (Blueprint $table) {
             $table->unsignedInteger('type_id');
             $table->unsignedInteger('attribute_id');
-            $table->float('value', 17, 4);
+            $table->unsignedInteger('value');
             $table->timestamps();
 
             $table->primary(['type_id', 'attribute_id']);
